+++ client/src/components/login/SignupLink/SignupLink.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import SignupLinkStyle from './SignupLink.style';
+
+function SignupLink ({ onLoginClose }) {
+  return (
+    <SignupLinkStyle 
+      onLoginClose={onLoginClose}
+    />
+  )
+}
+
+export default SignupLink;
