+++ client/src/components/login/SocialForm/SocialForm.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import SocialFormStyle from './SocialForm.style';
+
+function SocialForm() {
+  return (
+    <SocialFormStyle />
+  );
+}
+
+export default SocialForm;
